Refactor Docker execution to use background jobs
- Move Docker container execution logic from Agent model to RunJob
- Rename RunExecutionJob to simpler RunJob
- Task#run now queues a background job instead of blocking the request
- Prevents web requests from being blocked by long-running Docker containers
